notes: no to minimizing
If anything goes wrong, we'll want stack traces that point to source code. For
that, we'll want source maps, and for that, we'll have to switch to a webpack-
aware way of adding the shebang. It just so happens that webpack is broken
such that enabling the feature to add the shebang, namely BannerPlugin, also
breaks source map generation. Thus we won't be able to turn on minimize at
this time.