notes: no to 'waking up' socket
It turns out to have the same stuff as plain logs, and you can view those non-
application logs just fine even when you're not a member on the project. So
use `snail logs -ap <domain>` instead. The 'waking up' socket is also a lot
less friendly, with the program freezing up without any output when the
project is already awake.