pipe: send leading close parenthesis
We need some more decoration to make our protocol stand out from any
junk that comes out from the terminal session. The sad fact is that the
little pipe-wrap script is fallible. Errors being mixed into a single
stream with our data means that if anything goes wrong, our client can
misbehave.

I'm starting us off with `)`. It's a normal printable character in the
ASCII range, and it seems you wouldn't normally see it at the beginning
of a line because youd prefer to have it on the line before or to wrap a
whole word down with it.